abstracted FacilitiesList
